requirements.txt
# Core
boto3~=1.9.197
Django==2.2.4
gunicorn
pytz

# Django addons
django-countries~=5.4
django-environ~=0.4.5
django-hashid-field~=2.1.6
django-storages~=1.7.1

# Database
psycopg2-binary~=2.7.7

# Additional libs.
Pillow~=6.1.0
